moving llh_defs to utils, and simple_signal to data folders
