+++ pisa/stages/data/super_simple_pipeline.cfg
@@ -4,7 +4,7 @@ super_simple_binning.order = stuff
 super_simple_binning.stuff = {'bin_edges':np.linspace(0.,40.,51)}
 
 
-[simple_fit.pi_simple_signal]
+[data.pi_simple_signal]
 
 input_specs  = None
 calc_specs   = None
@@ -36,4 +36,4 @@ output_specs = super_simple_binning
 
 [pipeline]
 
-order = simple_fit.pi_simple_signal, likelihood.pi_generalized_llh_params
+order = data.pi_simple_signal, likelihood.pi_generalized_llh_params
